build: support build static library for linux
Target library name is rkcrypto_static because target
libraries file name can not be same in CMake.
So, `rkcrypto_static` should be add in link libraries
if you want to link static library.
Like this:
`target_link_libraries(librkcrypto_test rkcrypto_static)`

And the file name of `librkcrypto_static.a` will be output
to `librkcrypto.a` to keep the same with shared library.

Signed-off-by: Elon Zhang <zhangzj@rock-chips.com>
Change-Id: I2e9d6c74bd48511a3d21f4dad52291ce54c1c2c7